Rename diversity penalty to uniformity penalty
This doesn't change the user options, however it does affect the
output of the composite score since "diversity penalty" is renamed
into "uniformity penalty", in that sense it breaks backward
compatibility, but I think it is acceptable as the behavior otherwise
is unnaffected. This might only affect users who automatically parsed
the composite score with their own parsing code, which I think it
rather unlikely to exist.